# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Bazi Fortune App API (Go Primary)
  version: 1.0.0
  description: |
    Frozen API contract for the primary Go backend.
    Freeze date: 2026-02-20.
    Primary route source: backend/handlers/routes.go.
  x-contract-status: frozen
  x-freeze-date: "2026-02-20"
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production placeholder
...
paths:
...
  /api/v1/bazi/detail/{id}:
    get:
      tags: [Bazi]
      summary: Get bazi detail by record ID
      operationId: getBaziDetail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Bazi detail fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaziDetailResponse"
        "400":
          $ref: "#/components/responses/MixedError400"
        "401":
          $ref: "#/components/responses/MixedError401"
        "403":
          $ref: "#/components/responses/MixedError403"
        "404":
          $ref: "#/components/responses/MixedError404"
        "500":
          $ref: "#/components/responses/MixedError500"
  /api/v1/bazi:
    post:
      tags: [Bazi]
      summary: Create bazi record
      operationId: createBaziRecord
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BaziCalculateRequest"
      responses:
        "201":
          description: Bazi record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaziCalculateResponse"
        "400":
          $ref: "#/components/responses/MixedError400"
        "401":
          $ref: "#/components/responses/MixedError401"
        "500":
          $ref: "#/components/responses/MixedError500"
  /api/v1/bazi/{id}:
    put:
      tags: [Bazi]
      summary: Update bazi record
      operationId: updateBaziRecord
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BaziCalculateRequest"
      responses:
        "200":
          description: Bazi record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaziCalculateResponse"
        "400":
          $ref: "#/components/responses/MixedError400"
        "401":
          $ref: "#/components/responses/MixedError401"
        "403":
          $ref: "#/components/responses/MixedError403"
        "404":
          $ref: "#/components/responses/MixedError404"
        "500":
          $ref: "#/components/responses/MixedError500"

    delete:
      tags: [Bazi]
      summary: Delete bazi record
      operationId: deleteBaziRecord
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Bazi record deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeEnvelopeNoData"
        "400":
          $ref: "#/components/responses/MixedError400"
        "401":
          $ref: "#/components/responses/MixedError401"
        "403":
          $ref: "#/components/responses/MixedError403"
        "404":
          $ref: "#/components/responses/MixedError404"
        "500":
          $ref: "#/components/responses/MixedError500"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    MixedError400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MixedErrorResponse"
    MixedError401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MixedErrorResponse"
    MixedError403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MixedErrorResponse"
    MixedError404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MixedErrorResponse"
    MixedError500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MixedErrorResponse"
  schemas:
...
    BaziCalculateRequest:
      type: object
      required: [year, month, day, hour, minute, gender]
      properties:
        year:
          type: integer
          minimum: 1900
          maximum: 2100
        month:
          type: integer
          minimum: 1
          maximum: 12
        day:
          type: integer
          minimum: 1
          maximum: 31
        hour:
          type: integer
          minimum: 0
          maximum: 23
        minute:
          type: integer
          minimum: 0
          maximum: 59
        gender:
          type: string
          enum: [male, female]
        name:
          type: string
        timezone:
          type: string
          default: Asia/Shanghai
...
    BaziData:
      type: object
      description: Bazi record entity persisted in backend
      required: [id, user_id, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        chat_id:
          type: string
          format: uuid
          nullable: true
        input_data:
          oneOf:
            - $ref: "#/components/schemas/BaziInput"
            - type: string
          nullable: true
        result_data:
          oneOf:
            - $ref: "#/components/schemas/BaziResult"
            - type: string
          nullable: true
        analysis:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    BaziInput:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        day:
          type: integer
        hour:
          type: integer
        minute:
          type: integer
        gender:
          type: string
        name:
          type: string
        timezone:
          type: string
    BaziResult:
      type: object
      properties:
        year_pillar:
          type: string
        month_pillar:
          type: string
        day_pillar:
          type: string
        hour_pillar:
          type: string
        day_master:
          type: string
        zodiac:
          type: string
        nayin:
          type: string
        season:
          type: string
        solar_terms:
          type: array
          items:
            type: string
        lunar_date:
          type: string
        five_elements:
          type: object
          additionalProperties:
            type: integer
        shi_shen:
          type: object
          additionalProperties:
            type: string
        yong_shen:
          type: array
          items:
            type: string
        xi_shen:
          type: array
          items:
            type: string
        ji_shen:
          type: array
          items:
            type: string
        lucky_elements:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        numbers:
          type: array
          items:
            type: integer
        directions:
          type: array
          items:
            type: string
...
    CodeEnvelopeBase:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
    CodeEnvelopeError:
      allOf:
        - $ref: "#/components/schemas/CodeEnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                type: object
                additionalProperties: true
...
    CodeEnvelopeNoData:
      $ref: "#/components/schemas/CodeEnvelopeBase"
    BaziCalculateResponse:
      type: object
      required: [code, message, data, result]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/BaziData"
        result:
          $ref: "#/components/schemas/BaziResult"
    BaziDetailResponse:
      type: object
      required: [code, message, data, input, result]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/BaziData"
        input:
          $ref: "#/components/schemas/BaziInput"
        result:
          $ref: "#/components/schemas/BaziResult"
...
    LocalizedErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
        data:
          nullable: true
...
    BooleanErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
    MixedErrorResponse:
      oneOf:
        - $ref: "#/components/schemas/CodeEnvelopeError"
        - $ref: "#/components/schemas/LocalizedErrorResponse"
        - $ref: "#/components/schemas/BooleanErrorResponse"